Adjust micro post font sizes.
I've had a few people comment that the larger font size on micro post
detail pages looked odd to them. Now it's shrinked down to the same font
size as what is used in micro posts in post listings.

I like the idea of a consistent font size, and I like how micro posts
currently look in post listings, but I'm still not sure that I like this
font size for the detail pages. I'm concerned that very short posts will
be overpowered by other page elements, like the comment box.

We'll try it for a bit!